+++ storage/src/lib.rs
@@ -1,7 +1,8 @@
 pub mod common;
 pub mod compaction;
 pub mod iterators;
-pub mod lsm_storage;
+mod lsm_storage;
+mod lsm_storage_inner;
 mod lsm_util;
 mod manifest;
 pub mod memtable;
@@ -9,5 +10,6 @@ mod wal;
 
 mod block;
 mod sst;
-pub use lsm_storage::{Config, Storage, new};
+pub use lsm_storage::new;
+pub use lsm_storage_inner::Config;
 pub use sst::{FileObject, SSTable, SSTableBuilder, SSTableIterator};
